Release Java SDK 4.0.6:
잔액 타입 오류 수정 Int -> Float

일부 모델에 설명 추가

+++ src/main/java/net/nurigo/sdk/message/model/Balance.kt
@@ -7,10 +7,10 @@ data class Balance(
     /**
      * 충전금액
      */
-    var balance: Int? = null,
+    var balance: Float? = null,
 
     /**
      * 포인트(충전금액과 가치 동일)
      */
-    var point: Int? = null,
+    var point: Float? = null,
 )
